feat(hackathon): delete implemented
